Vali:  - A lot of fixes remaining to system startup again after the rewrite of libgracht and swap of toolchains.

+++ librt/libc/include/internal/_string.h
@@ -3,7 +3,7 @@
 
 /* Nonzero if either X or Y is not aligned on a "long" boundary.  */
 #define UNALIGNED(X, Y) \
-	(((long)X & (sizeof (long) - 1)) | ((long)Y & (sizeof (long) - 1)))
+	(((long)(intptr_t)(X) & (sizeof (long) - 1)) | ((long)(intptr_t)(Y) & (sizeof (long) - 1)))
 
 /* How many bytes are copied each iteration of the 4X unrolled loop.  */
 #define BIGBLOCKSIZE    (sizeof (long) << 2)
