+++ librt/libc/mem/memset.c
@@ -6,7 +6,7 @@
 #include <string.h>
 
 #define LBLOCKSIZE (sizeof(long))
-#define UNALIGNED(X)   ((long)X & (LBLOCKSIZE - 1))
+#define UNALIGNED(X)   ((long)(intptr_t)(X) & (LBLOCKSIZE - 1))
 #define TOO_SMALL(LEN) ((LEN) < LBLOCKSIZE)
 
 #if defined(_MSC_VER) && !defined(__clang__)
