+++ librt/libc/string/strncat.c
@@ -8,7 +8,7 @@
 
 /* Nonzero if X is aligned on a "long" boundary.  */
 #define ALIGNED(X) \
-	(((long)X & (sizeof (long) - 1)) == 0)
+	(((long)(intptr_t)(X) & (sizeof (long) - 1)) == 0)
 
 #if LONG_MAX == 2147483647L
 #define DETECTNULL(X) (((X) - 0x01010101) & ~(X) & 0x80808080)
